LightSheetControl renamed to PairedStageControl. Suggested by Jon Daniels and makes a lot of sense. Also fixed empty frame shown when no Multi Stage adapter was found.
