+++ plugins/PairedStageControl/build.xml
@@ -1,3 +1,3 @@
-<project name="LightSheetControl" default="jar">
+<project name="PairedStageControl" default="jar">
 	<import file="../javapluginbuild.xml"/>
 </project>
