+++ plugins/PairedStageControl/src/main/java/org/micromanager/pairedstagecontrol/ReferencePoint.java
@@ -2,7 +2,7 @@
  * @author kthorn
  */
 
-package org.micromanager.lightsheetcontrol;
+package org.micromanager.pairedstagecontrol;
 
 import java.io.Serializable;
